Optimize transfer chain selector ux
